Deduplicate GLSL location assignment helpers shared by SPIR-V and WebGPU backends.
