+++ CMakeLists.txt
@@ -220,7 +220,7 @@ file(GLOB PLATFORM_COMMON_FILES
 list(APPEND TGFX_FILES ${PLATFORM_COMMON_FILES})
 
 if (NOT TGFX_USE_METAL AND NOT TGFX_USE_VULKAN AND NOT TGFX_USE_D3D12)
-    file(GLOB SHADER_COMPILER_FILES src/gpu/ShaderCompiler.*)
+    file(GLOB SHADER_COMPILER_FILES src/gpu/ShaderCompiler.cpp)
     if (SHADER_COMPILER_FILES)
         list(REMOVE_ITEM TGFX_FILES ${SHADER_COMPILER_FILES})
     endif ()
