+++ src/gpu/ShaderCompiler.cpp
@@ -26,34 +26,6 @@
 
 namespace tgfx {
 
-// Replace a regex pattern in the source string by invoking a callback for each match.
-// The callback receives the match and returns the replacement string.
-using MatchReplacer = std::string (*)(const std::smatch&, int&);
-
-static std::string replaceAllMatches(const std::string& source, const std::regex& pattern,
-                                     MatchReplacer replacer, int& counter) {
-  std::smatch match;
-  std::string::const_iterator searchStart(source.cbegin());
-  std::string result;
-  size_t lastPos = 0;
-  while (std::regex_search(searchStart, source.cend(), match, pattern)) {
-    auto matchPos = static_cast<size_t>(match.position(0));
-    auto iterOffset = static_cast<size_t>(searchStart - source.cbegin());
-    size_t matchStart = matchPos + iterOffset;
-    result += source.substr(lastPos, matchStart - lastPos);
-    result += replacer(match, counter);
-    lastPos = matchStart + static_cast<size_t>(match.length(0));
-    searchStart = match.suffix().first;
-  }
-  result += source.substr(lastPos);
-  return result;
-}
-
-static std::string upgradeGLSLVersion(const std::string& source) {
-  static std::regex versionRegex(R"(#version\s+\d+(\s+es)?)");
-  return std::regex_replace(source, versionRegex, "#version 450");
-}
-
 // Assign fixed binding points for internal UBOs to match CPU-side constants:
 // VertexUniformBlock -> set 0, binding 0 (VERTEX_UBO_BINDING_POINT)
 // FragmentUniformBlock -> set 0, binding 1 (FRAGMENT_UBO_BINDING_POINT)
@@ -80,7 +52,7 @@ static std::string replaceCustomUBO(const std::smatch& match, int& counter) {
 static std::string assignCustomUBOBindings(const std::string& source) {
   static std::regex uboRegex(R"(layout\s*\(\s*std140\s*\)\s*uniform\s+(\w+))");
   int binding = 0;
-  return replaceAllMatches(source, uboRegex, replaceCustomUBO, binding);
+  return detail::ReplaceAllMatches(source, uboRegex, replaceCustomUBO, binding);
 }
 
 static std::string replaceSamplerBinding(const std::smatch& match, int& counter) {
@@ -94,163 +66,17 @@ static std::string replaceSamplerBinding(const std::smatch& match, int& counter)
 static std::string assignSamplerBindings(const std::string& source) {
   static std::regex samplerRegex(R"(uniform\s+(sampler\w+)\s+(\w+);)");
   int binding = 0;
-  return replaceAllMatches(source, samplerRegex, replaceSamplerBinding, binding);
-}
-
-static std::string replaceInputLocation(const std::smatch& match, int& counter) {
-  std::string interpStr = match[1].matched ? match[1].str() : "";
-  std::string precisionStr = match[2].matched ? match[2].str() : "";
-  // match[5] is the whole array suffix like "[5]" (preserved verbatim in the emitted decl);
-  // match[6] is just the size digits, used to advance the location counter.
-  std::string arraySuffix = match[5].matched ? match[5].str() : "";
-  int step = match[6].matched ? std::stoi(match[6].str()) : 1;
-  std::string decl = "layout(location=" + std::to_string(counter) + ") " + interpStr + "in " +
-                     precisionStr + match[3].str() + " " + match[4].str() + arraySuffix + ";";
-  counter += step;
-  return decl;
-}
-
-// Rewrite interface-variable declarations so their `layout(location=N)` numbers are assigned by
-// looking up each variable's name in `nameToLocation`, instead of in source order. The map is
-// built from the same regex scan this pass consumes, so every matched name is guaranteed to be
-// present; the assertion below documents that contract.
-static std::string replaceWithNameKeyedLocations(
-    const std::string& source, const std::regex& pattern,
-    const std::unordered_map<std::string, int>& nameToLocation, bool isInput) {
-  std::smatch match;
-  std::string::const_iterator searchStart(source.cbegin());
-  std::string result;
-  size_t lastPos = 0;
-  while (std::regex_search(searchStart, source.cend(), match, pattern)) {
-    auto matchPos = static_cast<size_t>(match.position(0));
-    auto iterOffset = static_cast<size_t>(searchStart - source.cbegin());
-    size_t matchStart = matchPos + iterOffset;
-    result += source.substr(lastPos, matchStart - lastPos);
-
-    std::string interpStr = match[1].matched ? match[1].str() : "";
-    std::string precisionStr = match[2].matched ? match[2].str() : "";
-    std::string arraySuffix = match[5].matched ? match[5].str() : "";
-    const std::string& name = match[4].str();
-    auto it = nameToLocation.find(name);
-    DEBUG_ASSERT(it != nameToLocation.end());
-    int location = it->second;
-    std::string decl = "layout(location=" + std::to_string(location) + ") " + interpStr +
-                       (isInput ? "in " : "out ") + precisionStr + match[3].str() + " " + name +
-                       arraySuffix + ";";
-    result += decl;
-
-    lastPos = matchStart + static_cast<size_t>(match.length(0));
-    searchStart = match.suffix().first;
-  }
-  result += source.substr(lastPos);
-  return result;
-}
-
-struct NameSortedEntry {
-  std::string name;
-  int step;
-};
-
-static bool CompareEntryNames(const NameSortedEntry& a, const NameSortedEntry& b) {
-  return a.name < b.name;
-}
-
-// Build a `name -> location` map from the interface declarations matched by `pattern`. Names are
-// collected in source order, then sorted lexicographically before locations are assigned. Array-
-// typed varyings occupy `size` consecutive locations, so the counter is advanced by the array
-// length to match SPIR-V semantics for per-vertex block members.
-static std::unordered_map<std::string, int> buildNameSortedLocationMap(const std::string& source,
-                                                                       const std::regex& pattern) {
-  std::vector<NameSortedEntry> entries;
-  std::smatch match;
-  std::string::const_iterator searchStart(source.cbegin());
-  while (std::regex_search(searchStart, source.cend(), match, pattern)) {
-    int step = match[6].matched ? std::stoi(match[6].str()) : 1;
-    entries.push_back({match[4].str(), step});
-    searchStart = match.suffix().first;
-  }
-  std::sort(entries.begin(), entries.end(), CompareEntryNames);
-  std::unordered_map<std::string, int> nameToLocation;
-  int location = 0;
-  for (const auto& entry : entries) {
-    nameToLocation[entry.name] = location;
-    location += entry.step;
-  }
-  return nameToLocation;
-}
-
-// Add location qualifiers to 'in' variables, handling optional interpolation qualifiers
-// (flat, noperspective), precision qualifiers (highp, mediump, lowp), and array-typed varyings
-// like `in vec2 blurCoordinates[5];`. Array-typed varyings occupy `size` consecutive locations,
-// so the counter is advanced by the array length instead of one, matching SPIR-V semantics for
-// per-vertex block members.
-//
-// Vertex stage inputs are vertex attributes whose locations must match the attribute order
-// declared in `RenderPipelineDescriptor::vertex.bufferLayouts` (a CPU-side contract shared by
-// Metal / Vulkan / D3D12), so we keep the source-declaration order there. Fragment stage inputs
-// are varyings paired with the vertex stage's outputs; assigning their locations by sorted name
-// makes the pairing independent of source order, so a fragment shader can declare its `in`
-// varyings in any order and still line up with the vertex `out` declarations.
-static std::string assignInputLocationQualifiers(const std::string& source, ShaderStage stage) {
-  static std::regex inVarRegex(
-      R"((flat\s+|noperspective\s+)?in\s+(highp\s+|mediump\s+|lowp\s+)?(\w+)\s+(\w+)(\s*\[\s*(\d+)\s*\])?\s*;)");
-  if (stage == ShaderStage::Vertex) {
-    int location = 0;
-    return replaceAllMatches(source, inVarRegex, replaceInputLocation, location);
-  }
-  auto nameToLocation = buildNameSortedLocationMap(source, inVarRegex);
-  return replaceWithNameKeyedLocations(source, inVarRegex, nameToLocation, /*isInput=*/true);
-}
-
-static std::string replaceOutputLocation(const std::smatch& match, int& counter) {
-  std::string interpStr = match[1].matched ? match[1].str() : "";
-  std::string precisionStr = match[2].matched ? match[2].str() : "";
-  std::string arraySuffix = match[5].matched ? match[5].str() : "";
-  int step = match[6].matched ? std::stoi(match[6].str()) : 1;
-  std::string decl = "layout(location=" + std::to_string(counter) + ") " + interpStr + "out " +
-                     precisionStr + match[3].str() + " " + match[4].str() + arraySuffix + ";";
-  counter += step;
-  return decl;
-}
-
-// Add location qualifiers to 'out' variables, handling optional interpolation qualifiers
-// (flat, noperspective), precision qualifiers (highp, mediump, lowp), and array-typed varyings
-// like `out vec2 blurCoordinates[5];`. Array-typed varyings occupy `size` consecutive locations,
-// so the counter is advanced by the array length instead of one, matching SPIR-V semantics for
-// per-vertex block members.
-//
-// Vertex stage outputs are varyings feeding the fragment stage; assigning their locations by
-// sorted name (the same rule used for fragment inputs) makes the vertex→fragment pairing
-// independent of source order. Fragment stage outputs are colour attachments whose location
-// index maps directly to the colour attachment slot, so we keep the source-declaration order.
-// Name-based pairing assumes both stages declare the same set of interface names: a varying
-// emitted by the vertex shader but never consumed by the fragment (valid GLSL on the OpenGL
-// backend) shifts the vertex-side locations and desyncs the two stages.
-static std::string assignOutputLocationQualifiers(const std::string& source, ShaderStage stage) {
-  static std::regex outVarRegex(
-      R"((flat\s+|noperspective\s+)?out\s+(highp\s+|mediump\s+|lowp\s+)?(\w+)\s+(\w+)(\s*\[\s*(\d+)\s*\])?\s*;)");
-  if (stage == ShaderStage::Fragment) {
-    int location = 0;
-    return replaceAllMatches(source, outVarRegex, replaceOutputLocation, location);
-  }
-  auto nameToLocation = buildNameSortedLocationMap(source, outVarRegex);
-  return replaceWithNameKeyedLocations(source, outVarRegex, nameToLocation, /*isInput=*/false);
-}
-
-// Remove precision qualifiers that are not supported in desktop GLSL 450.
-static std::string removePrecisionDeclarations(const std::string& source) {
-  static std::regex precisionDeclRegex(R"(precision\s+(highp|mediump|lowp)\s+\w+\s*;)");
-  return std::regex_replace(source, precisionDeclRegex, "");
+  return detail::ReplaceAllMatches(source, samplerRegex, replaceSamplerBinding, binding);
 }
 
 std::string PreprocessGLSL(const std::string& glslCode, ShaderStage stage) {
-  auto result = upgradeGLSLVersion(glslCode);
+  auto result = detail::UpgradeGLSLVersion(glslCode);
   result = assignInternalUBOBindings(result);
   result = assignCustomUBOBindings(result);
   result = assignSamplerBindings(result);
-  result = assignInputLocationQualifiers(result, stage);
-  result = assignOutputLocationQualifiers(result, stage);
-  result = removePrecisionDeclarations(result);
+  result = detail::AssignInputLocationQualifiers(result, stage);
+  result = detail::AssignOutputLocationQualifiers(result, stage);
+  result = detail::RemovePrecisionDeclarations(result);
   return result;
 }
